requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.4
django-appconf==1.0.2
django-filter==1.1.0
django-imagekit==4.0.1
django-material==1.1.1
djangorestframework==3.7.1
get==0.0.39
Markdown==2.6.9
pilkit==2.0
post==0.0.26
psycopg2==2.7.3.2
public==0.0.65
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.2
query-string==0.0.28
router==0.1
six==1.10.0
static3==0.7.0
